backend/python_scripts/requirements.txt
pandas
pymongo
openpyxl
python-dotenv
